+++ src/model/user.rs
@@ -30,7 +30,7 @@ pub struct Info {
 
     pub avatar_key: Option<String>,
     pub avatar_uploaded: bool,
-    pub avatar_version: i64,
+    pub avatar_version: u32,
 
     pub is_sadmin: bool,
 
@@ -60,7 +60,7 @@ pub struct Form {
 pub struct AvatarReservation {
     pub object_key: String,
     pub prev_object_key: Option<String>,
-    pub avatar_version: i64,
+    pub avatar_version: u32,
 }
 
 /// A stored password credential used during login verification.
